python/requirements.txt
websocket-client
numpy
mss
argparse